requirements.txt
# FastAPI Auth Service
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
redis>=5.0.1
fastapi-limiter>=0.1.6
python-jose[cryptography]>=3.3.0
pydantic[email]>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6

# Auth package (Argon2 hashing)
argon2-cffi>=23.1.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
httpx>=0.26.0

# Development
black>=24.1.0
ruff>=0.1.14
